refactor(studio): extract ExperimentGroupCard and UpdatedAt into own files
- Move ExperimentGroupCard to ExperimentGroupCard.tsx (one component per file)
- Move UpdatedAt to UpdatedAt.tsx
- Replace arbitrary gap values with Tailwind scale (gap-2, gap-1.5)
